V1.0.1/build/msvc (#156)
* Refactor: Added WinAPI patch instead of manually fixing issues each time

* Refactor: Added WinAPI patch instead of manually fixing issues each time

* Refactor: Changed macro concatination to use separate macro SENC_CAT

* Refactor: Added custom primitive sampling to wrap cannon sampling

* Doc: Added file doc for winapi_patch.hpp

* Refactor: Replaced experimental std::scope_exit with custom utils::AtScopeExit

* Fix: Fixed bad parentheses on call

* Fix: Removed bad const

* Refactor: Wrapped concept macros with a seprate SENC_EXPAND macro

* Refactor: Changed alias Port from int to unsigned short

* Fix: Added missing include

* Refactor: added explicit static cast

* Refactor: Removed unused vars

* Refactor: Added explicit cast of mt19937 seed

* Refactor: Changed Port alias to uint16_t and added parse_port

* Refactor: Added explicit static cast

* Refactor: Changed input type for bigint inputs from long long to long

* Refactor: Added const ref

* Fix: Removed bad noexcept from UUID generate() overloads

* Refactor: Replaced UUID generate() overloads with separate methods with separate names

* Build: Added MSVC configuration

* Fix: Separated MSVC concept macro definitions from non-MSVC ones to make code properly compile in both MSVC and gcc

* Chore: Added direct include to env.hpp

+++ senc/utils/Fraction_impl.hpp
@@ -165,7 +165,7 @@ namespace senc::utils
 			this->_num = -this->_num;
 		}
 
-		long long g = std::gcd(this->_num, this->_den);
+		Int g = static_cast<Int>(std::gcd(this->_num, this->_den));
 		this->_num /= g;
 		this->_den /= g;
 	}
